requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-multipart
psycopg2
pytz
requests
fastavro
pandas